Website/backend/requirements.txt
attrs==21.4.0
backend==0.2.4.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
Flask==2.1.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
marshmallow==3.15.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.8
pytest==7.1.1
six==1.16.0
tomli==2.0.1
Unidecode==1.3.4
urllib3==1.26.9
webargs==8.1.0
Werkzeug==2.1.1
zipp==3.8.0
pandas~=1.4.2
setuptools~=62.1.0
xlwings~=0.27.6
pyxlsb~=1.0.9
unicode~=2.8
Pillow~=9.1.0